+++ src/main/java/models/goods/GoodsType.java
@@ -1,4 +1,4 @@
-package models;
+package models.goods;
 
 import java.awt.print.Paper;
