+++ src/main/java/models/rollingstock/CarType.java
@@ -1,4 +1,4 @@
-package models;
+package models.rollingstock;
 
 public enum CarType
 {
